backend/requirements.txt
fastapi==0.122.0
protobuf==6.33.1
pydantic==2.12.4
python-dotenv==1.2.1
strictjson==6.3.0
